requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
boto3==1.40.40
botocore==1.40.40
cachetools==5.5.0
capstone==6.0.0a4
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.21.1
cloudpickle==3.1.2
colorama==0.4.6
colored-traceback==0.4.2
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
datasets==2.19.1
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.1.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
fastcoref==2.1.6
fastjsonschema==2.20.0
filelock==3.20.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.2.4
groq==1.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.7.1
httpx==0.27.2
huggingface-hub==0.35.3
idna==3.10
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.46.0
Mako==1.3.10
marisa-trie==1.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
murmurhash==1.0.13
narwhals==1.14.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netcat==0.0.1a0
networkx==3.5
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.63.1
numpy==2.1.3
orjson==3.11.3
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
pefile==2023.2.7
pillow==11.0.0
platformdirs==4.3.6
plumbum==1.9.0
preshed==3.0.10
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.4.1
protobuf==5.29.0
psutil==6.1.0
pure_eval==0.2.3
pwntools==4.14.1
pyarrow==21.0.0
pyarrow-hotfix==0.7
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
pyelftools==0.32
Pygments==2.18.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyNaCl==1.5.0
pyparsing==3.2.0
pyserial==3.5
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2025.7.34
reportlab==4.4.7
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.15.1
rignore==0.7.0
ROPGadget==7.6
rpds-py==0.21.0
rpyc==6.0.2
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.1.2
sentry-sdk==2.41.0
setuptools==75.5.0
shap==0.50.0
shellingham==1.5.4
shiboken6==6.9.1
six==1.16.0
slicer==0.0.8
smart_open==7.3.0.post1
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
-e git+https://github.com/entbappy/Movie-Recommender-System-Using-Machine-Learning@ae06ef2abddf9daacdaf587f335228a494fddffb#egg=src
srsly==2.5.1
stack-data==0.6.3
starlette==0.48.0
streamlit==1.40.2
sympy==1.14.0
tenacity==9.0.0
terminado==0.18.1
thinc==8.3.6
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.16.1
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.11.0
unicorn==2.1.2
unix-ar==0.2.1
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.37.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.3
xgboost==3.0.5
xxhash==3.6.0
yarl==1.22.0
zstandard==0.23.0
